Bug 1990586 - Make `WhiteSpaceVisibilityKeeper::DeleteContentNodeAndJoinTextNodesAroundIt` track given caret point r=m_kato
It's used as the result in some cases.  However, it's not tracked in
some cases when touching the DOM.

This modifies the API of `AutoTrackDOMPoint` to flush the tracking point
but can keep tracking.  We should use the new API in every place later.

Differential Revision: https://phabricator.services.mozilla.com/D266752

UltraBlame original commit: 6fafb4a41c0660db7d6fd5b7790c267615028036