Bug 1991761 - Pass architecture arg to Translations e2e tests r=translations-reviewers,gregtatum
This patch updates the Translations end-to-end test arguments
to take an argument for the model architecture to test.

This is a partial patch that will not work as written,
but will work with the subsequent changes in the stack.

Differential Revision: https://phabricator.services.mozilla.com/D266892

UltraBlame original commit: cab202f70fa90acb3110c6a90a06e80edbdef0d6