Bug 1973820 p4 - Fix Nimbus test which was relying on state from other tests r=cpeterson
The Nimbus test was accidentally relying on the prior state of
`foo.enabled` in one assertion. There was also another separate part of
the test which I've distinguished with a comment for clarity.

This patch stack changes gtests to run as separate processes per test
suite, so since the `foo.enabled` value was being set in the
`NimbusFeaturesGet` test suite and the assertion in question is in the
`NimbusFeaturesUpdate` test suite, the assertion now fails. For what
it's worth, it might make more sense for all of these tests to be under
a single `NimbusFeatures` suite, however either way the tests should not
rely on each other.

Differential Revision: https://phabricator.services.mozilla.com/D255176

UltraBlame original commit: f6940eb64370414a8cdb5a671c7b8737cd59149c