Bug 1991240 - Restore unified toolbar appearance in macOS Tahoe. r=desktop-theme-reviewers,emilio
In Tahoe, the titlebar material is no longer used on the titlebar itself, rather, the window background material is used. Therefore, no titlebar material needs to be drawn in the toolbar.

The titlebar separator is removed in Tahoe by making the titlebar transparent, as NSTitlebarSeparatorStyleNone no longer removes it. Due to the above, this does not change the titlebar appearance other than removing the separator.

Differential Revision: https://phabricator.services.mozilla.com/D267166

UltraBlame original commit: ed1483c9c4f241856e588ee93c17906710679e85