Bug 1992422 - part 2: Convert GetRowSpec and GetColSpec to Span. r=layout-reviewers,emilio
Nobody ever checks the failure value, so return the Span directly.

This drops the setting of mNumRows and mNumCols in MouseDrag(), because we
already require that the number of rows or columns does not change. This is
checked via an assertion in GenerateRowCol().

Differential Revision: https://phabricator.services.mozilla.com/D267673

UltraBlame original commit: 7d6365c959e857e3a6b47c17d836a6338205d1ef