Bug 1778684 - Add test coverage for blocklisted addon r=rpl
When bug 1917848 introduced the blocklist messages in the extensions
panel, it added logic to customize the about:addons page that would be
opened if there are no extensions in the panel. It did not include unit
tests for this scenario.

The previous commit in this bug (bug 1778684) defaults to showing the
panel when extensions are disabled, with some explanatory UI. Because
blocklisted add-ons are disabled, this also implies that the
shouldShowBlocklistAttention check in browser-addons.js is now obsolete.
This patch removes that check as a cleanup.

Differential Revision: https://phabricator.services.mozilla.com/D265968

UltraBlame original commit: df293fdfa6ae20a762890f58815d66e977563aba