Bug 1992014: Allocate StackMaps using LifoAlloc, deduplicate. r=rhunt
Stack maps were individually malloced, which produced significant
amounts of lock contention and cost a significant amount of time in
baseline compiles.

This patch updates the StackMaps class to use a LifoAlloc to allocate
stack maps. The LifoAlloc will malloc chunks of memory to fill with
stack maps; after being added to the LifoAlloc, the pointer to the stack
map is added to the already-existing hash map and remains stable.

Stack maps are also now deduplicated on add. They are compared to the
most recently added stack map, and if they are identical, the existing
stack map is used again, shrinking the number of stack maps we have to
store.

Differential Revision: https://phabricator.services.mozilla.com/D267110

UltraBlame original commit: e25053debed5d2dbe4a7d15c65b210f701f3414c